+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -1,12 +1,16 @@
 # Version history for `cardano-ledger-core`
 
-## 1.0.1.0
+## 1.1.0.0
 
+* Add `ToJSON` superclass constraints for `EraTxOut`.
+* Add superclass constraints for `Val`:
+  * `NoThunks`, `EncCBOR`, `DecCBOR`, `ToJSON`, `NFData`, `Show`
+  * `EncCBOR (CompactForm t)`, `DecCBOR (CompactForm t)`
 * Add `ToJSON`/`FromJSON` instances for `CompactForm Coin`, `SafeHash` and `TxId`
 * Add `ToJSON` instances for:
   * `Ptr`, `CertIx`, `TxIx`
   * `Trip` and `UMap`
-  * `DeltaCoin`
+  * `DeltaCoin` and `CompactForm DeltaCoin`
   * `InstantaneousRewards`, `FutureGenDeleg`, `PState`, `DState` and `DPState`.
   * `UTxO` and `TxIn`